% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/location.R
\docType{class}
\name{location-class}
\alias{location-class}
\title{An S4 Class to represent a location object.}
\description{
An S4 Class to represent a location object.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character vector}

\item{\code{street}}{A character vector}

\item{\code{zip_code}}{A character vector}

\item{\code{community}}{A character vector}

\item{\code{canton}}{A character vector of length 2}

\item{\code{lat}}{A numeric value}

\item{\code{lng}}{A numeric value}
}}

